refactor: move buildSchema into utils file
It avoid circular dependencies and it is temporary
It will be moved again in #35158
