+++ packages/react-native-codegen/src/parsers/flow/index.js
@@ -11,71 +11,12 @@
 'use strict';
 
 import type {SchemaType} from '../../CodegenSchema.js';
-import type {Parser} from '../parser';
 
-// $FlowFixMe[untyped-import] there's no flowtype flow-parser
-const flowParser = require('flow-parser');
 const fs = require('fs');
-const {
-  buildSchemaFromConfigType,
-  getConfigType,
-  isModuleRegistryCall,
-} = require('../utils');
-const {buildComponentSchema} = require('./components');
-const {wrapComponentSchema} = require('./components/schema');
-const {buildModuleSchema} = require('./modules');
+const {buildSchema} = require('./utils');
 const {FlowParser} = require('./parser');
 
 const parser = new FlowParser();
-
-function Visitor(infoMap: {isComponent: boolean, isModule: boolean}) {
-  return {
-    CallExpression(node: $FlowFixMe) {
-      if (
-        node.callee.type === 'Identifier' &&
-        node.callee.name === 'codegenNativeComponent'
-      ) {
-        infoMap.isComponent = true;
-      }
-
-      if (isModuleRegistryCall(node)) {
-        infoMap.isModule = true;
-      }
-    },
-    InterfaceExtends(node: $FlowFixMe) {
-      if (node.id.name === 'TurboModule') {
-        infoMap.isModule = true;
-      }
-    },
-  };
-}
-
-function buildSchema(
-  contents: string,
-  filename: ?string,
-  parser: Parser,
-): SchemaType {
-  // Early return for non-Spec JavaScript files
-  if (
-    !contents.includes('codegenNativeComponent') &&
-    !contents.includes('TurboModule')
-  ) {
-    return {modules: {}};
-  }
-
-  const ast = flowParser.parse(contents, {enums: true});
-  const configType = getConfigType(ast, Visitor);
-
-  return buildSchemaFromConfigType(
-    configType,
-    filename,
-    ast,
-    wrapComponentSchema,
-    buildComponentSchema,
-    buildModuleSchema,
-    parser,
-  );
-}
 
 function parseModuleFixture(filename: string): SchemaType {
   const contents = fs.readFileSync(filename, 'utf8');
@@ -88,7 +29,6 @@ function parseString(contents: string, filename: ?string): SchemaType {
 }
 
 module.exports = {
-  buildSchema,
   parseModuleFixture,
   parseString,
 };
