+++ packages/react-native-codegen/src/parsers/typescript/utils.js
@@ -11,13 +11,83 @@
 'use strict';
 
 import type {TypeAliasResolutionStatus, TypeDeclarationMap} from '../utils';
+import type {Parser} from '../parser';
 
+// $FlowFixMe[untyped-import] Use flow-types for @babel/parser
+const babelParser = require('@babel/parser');
+const {
+  buildSchemaFromConfigType,
+  getConfigType,
+  isModuleRegistryCall,
+} = require('../utils');
+const {buildComponentSchema} = require('./components');
+const {wrapComponentSchema} = require('./components/schema');
+const {buildModuleSchema} = require('./modules');
 const {parseTopLevelType} = require('./parseTopLevelType');
+import type {SchemaType} from '../../CodegenSchema';
+
+function Visitor(infoMap: {isComponent: boolean, isModule: boolean}) {
+  return {
+    CallExpression(node: $FlowFixMe) {
+      if (
+        node.callee.type === 'Identifier' &&
+        node.callee.name === 'codegenNativeComponent'
+      ) {
+        infoMap.isComponent = true;
+      }
+
+      if (isModuleRegistryCall(node)) {
+        infoMap.isModule = true;
+      }
+    },
+
+    TSInterfaceDeclaration(node: $FlowFixMe) {
+      if (
+        Array.isArray(node.extends) &&
+        node.extends.some(
+          extension => extension.expression.name === 'TurboModule',
+        )
+      ) {
+        infoMap.isModule = true;
+      }
+    },
+  };
+}
+
+function buildSchema(
+  contents: string,
+  filename: ?string,
+  parser: Parser,
+): SchemaType {
+  // Early return for non-Spec JavaScript files
+  if (
+    !contents.includes('codegenNativeComponent') &&
+    !contents.includes('TurboModule')
+  ) {
+    return {modules: {}};
+  }
+
+  const ast = babelParser.parse(contents, {
+    sourceType: 'module',
+    plugins: ['typescript'],
+  }).program;
+
+  const configType = getConfigType(ast, Visitor);
+
+  return buildSchemaFromConfigType(
+    configType,
+    filename,
+    ast,
+    wrapComponentSchema,
+    buildComponentSchema,
+    buildModuleSchema,
+    parser,
+  );
+}
 
 /**
  * TODO(T108222691): Use flow-types for @babel/parser
  */
-
 function getTypes(ast: $FlowFixMe): TypeDeclarationMap {
   return ast.body.reduce((types, node) => {
     switch (node.type) {
@@ -110,6 +180,7 @@ function resolveTypeAnnotation(
 }
 
 module.exports = {
+  buildSchema,
   resolveTypeAnnotation,
   getTypes,
 };
